+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added 🚀
 
+-   Added regression test
+
 ### Changed
 
 ### Removed 🗑
 
 ### Fixed 🐞
 
+-   Buildings are flattened when delta is active #824. 
 -   Selected Building now stays highlighted when map is rotated #1498
 
 ### Chore 👨‍💻 👩‍💻